// src/app/interfaces/token.interface.ts
export interface TokenPair {
  tokenAddress: string;
  pairAddress: string;
  exchange: string;
  exchangeLogo: string;
  price: number;
  marketCap: number;
  fullyDilutedValue: number;
  liquidity: number;
  volume24h: number;
  priceChange1h: number;
  priceChange24h: number;
  tokenIcon: string;
  tokenName: string;
  tokenSymbol: string;
} 